blogging/requirements.txt
Babel==2.7.0
Django==2.2.4
django-phonenumber-field==3.0.1
django-tinymce==2.8.0
phonenumbers==8.10.17
Pillow==6.1.0
pytz==2019.2
sqlparse==0.3.0
